requirements.txt
Django==1.11
django-annoying==0.10.4
django-appconf==1.0.2
django-avatar==4.0.1
django-bootstrap3==9.1.0
django-braces==1.12.0
django-crispy-forms==1.7.0
django-currentuser==0.2.0
django-datetime-widget==0.9.3
django-extensions==1.9.8
django-extra-views==0.9.0
django-filter==1.1.0
django-filters==0.2.1
django-simple-history==1.9.0
django-tables2==1.16.0
django-widget-tweaks==1.4.1
django-mptt
django-select2
djangorestframework
num2words
tablib
reportlab
pyserial
pyreportjasper


